+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix error passing `data` option to `Cookie` constructor
+
 1.16.0 / 2019-04-10
 ===================